feat: implement database API for saving, loading, and deleting presets and drafts
- Added `database.ts` for API interactions with the server to manage saved presets and drafts.
- Introduced functions to read, write, and delete saved presets.
- Implemented draft loading and saving functionality.
- Updated `main.ts` to utilize the new database functions for managing presets and drafts.

feat: enhance display settings and UI interactions

- Created `display-settings.ts` to manage theme settings and responsive UI behavior.
- Improved event handling for timeline generation input, allowing for smoother scrubbing and value updates.
- Added fullscreen toggle functionality with visual feedback.

fix: adjust growth transition logic for playback

- Modified `growth-transition.ts` to allow for non-eased updates during playback, improving animation consistency.

fix: update validation logic for generation parameters

- Adjusted `validateParams` to accept fractional generations for specific growth models.

style: enhance CSS for improved UI responsiveness and aesthetics

- Updated styles in `style.css` to support new UI components and improve overall layout and design.
- Added media queries for better mobile responsiveness.

test: add unit tests for growth transition and state validation

- Implemented tests for growth transition playback consistency.
- Added tests for validating generation parameters, including fractional values.

+++ backend/examples/benchmark.rs
@@ -6,18 +6,18 @@ use std::{hint::black_box, time::Instant};
 fn main() {
     let mut cases = Vec::new();
     let mut p: Plant = serde_json::from_value(presets()[0]["params"].clone()).unwrap();
-    p.generations = 6;
+    p.generations = 6 as f64;
     cases.push(("birch-6", p.clone()));
     p.premise = "F".into();
     p.rules = vec![Rule {
         expression: "F=F[+F][-F][&F]".into(),
     }];
-    p.generations = 7;
+    p.generations = 7 as f64;
     cases.push(("dense-7", p));
     for (name, p) in cases {
         let rules = engine::parse_rules(&p.rules).unwrap();
         let work = || {
-            let s = engine::expand(&p.premise, &rules, p.generations).unwrap();
+            let s = engine::expand(&p.premise, &rules, p.generations as u32).unwrap();
             let data = engine::geometry(&s, &p).unwrap();
             black_box((s, data))
         };
